8.2.20. добавляет отдельный компонент для кнопки Избранное.
