+++ project/src/components/offer-card-detal/offer-card-detal.tsx
@@ -1,15 +1,16 @@
 import {Link} from 'react-router-dom';
 import {Offer} from '../../types/offer';
 import {getRating} from '../../utils';
-import {AppRoute, AuthorizationStatus} from '../../const';
+import {AppRoute} from '../../const';
 import {Navigate} from 'react-router-dom';
 import {useState} from 'react';
-import {useAppSelector, useAppDispatch} from '../../hooks/index';
-import {getAuthorizationStatus} from '../../store/user-process/selectors';
-import {changeFavoriteStatusAction} from '../../store/api-actions';
-import { getOffers, getFavoritesOffers, getOffersNearby } from '../../store/data-process/selectors';
-import {updateOffersAndFavoritesOffers} from '../../store/data-process/update-data';
-import {updateOffers, updateFavoritesOffers, updateOffersNearby} from '../../store/data-process/data-process';
+// import {useAppSelector} from '../../hooks/index';
+// import {getAuthorizationStatus} from '../../store/user-process/selectors';
+// import {changeFavoriteStatusAction} from '../../store/api-actions';
+// import { getOffers, getFavoritesOffers, getOffersNearby } from '../../store/data-process/selectors';
+// import {updateOffersAndFavoritesOffers} from '../../store/data-process/update-data';
+// import {updateOffers, updateFavoritesOffers, updateOffersNearby} from '../../store/data-process/data-process';
+import {OfferCardDetalButton} from '../offer-card-detal-button/offer-card-detal-button';
 
 
 type OfferCardDetalProps = {
@@ -23,13 +24,13 @@ export function OfferCardDetal(props: OfferCardDetalProps): JSX.Element{
   const { offer, fromOfferPage, fromFavoritePage} = props;
 
   const [isNavigationOffer, setNavigationOffer] = useState(false);
-  const [isNavigationLogin_, setNavigationLogin_] = useState(false);
+  // const [isNavigationLogin_, setNavigationLogin_] = useState(false);
 
-  const authorizationStatus = useAppSelector(getAuthorizationStatus);
-  const offers = useAppSelector(getOffers);
-  const favoritesOffers = useAppSelector(getFavoritesOffers);
-  const offersNearby = useAppSelector(getOffersNearby);
-  const dispatch = useAppDispatch();
+  // const authorizationStatus = useAppSelector(getAuthorizationStatus);
+  // const offers = useAppSelector(getOffers);
+  // const favoritesOffers = useAppSelector(getFavoritesOffers);
+  // const offersNearby = useAppSelector(getOffersNearby);
+  // const dispatch = useAppDispatch();
 
   const {
     id,
@@ -47,34 +48,34 @@ export function OfferCardDetal(props: OfferCardDetalProps): JSX.Element{
     return <Navigate to={AppRoute.Offer + String(id)} />;
   }
 
-  if (isNavigationLogin_ && authorizationStatus.status !== AuthorizationStatus.Auth) {
-    return <Navigate to={AppRoute.Login} />;
-  }
+  // if (isNavigationLogin_ && authorizationStatus.status !== AuthorizationStatus.Auth) {
+  //   return <Navigate to={AppRoute.Login} />;
+  // }
 
-  const updateData = (update: Offer) => {
-    const result = updateOffersAndFavoritesOffers(id, update, offers, favoritesOffers, offersNearby);
-    if(result.offers) {
-      dispatch(updateOffers(result.offers));
-    }
-    if(result.favoritesOffers) {
-      dispatch(updateFavoritesOffers(result.favoritesOffers));
-    }
-    if(result.offersNearby) {
-      dispatch(updateOffersNearby(result.offersNearby));
-    }
-  };
+  // const updateData = (update: Offer) => {
+  //   const result = updateOffersAndFavoritesOffers(id, update, offers, favoritesOffers, offersNearby);
+  //   if(result.offers) {
+  //     dispatch(updateOffers(result.offers));
+  //   }
+  //   if(result.favoritesOffers) {
+  //     dispatch(updateFavoritesOffers(result.favoritesOffers));
+  //   }
+  //   if(result.offersNearby) {
+  //     dispatch(updateOffersNearby(result.offersNearby));
+  //   }
+  // };
 
-  const handleFavoriteStatusClick = () => {
-    setNavigationLogin_(true);
-    if(authorizationStatus.status === AuthorizationStatus.Auth) {
-      const statusId = {
-        id: String(id),
-        status: isFavorite ? '0' : '1',
-        updateData,
-      };
-      dispatch(changeFavoriteStatusAction(statusId));
-    }
-  };
+  // const handleFavoriteStatusClick = () => {
+  //   setNavigationLogin_(true);
+  //   if(authorizationStatus.status === AuthorizationStatus.Auth) {
+  //     const statusId = {
+  //       id: String(id),
+  //       status: isFavorite ? '0' : '1',
+  //       updateData,
+  //     };
+  //     dispatch(changeFavoriteStatusAction(statusId));
+  //   }
+  // };
 
   const handleCardClick = () => {
     setNavigationOffer(true);
@@ -88,20 +89,9 @@ export function OfferCardDetal(props: OfferCardDetalProps): JSX.Element{
           <b className="place-card__price-value">&euro;{price}</b>
           <span className="place-card__price-text">&#47;&nbsp;night</span>
         </div>
-        <button
-          className={`place-card__bookmark-button button ${
-            isFavorite && 'place-card__bookmark-button--active'
-          }`}
-          type="button"
-          onClick={handleFavoriteStatusClick}
-        >
-          <svg className="place-card__bookmark-icon" width="18" height="19">
-            <use xlinkHref="#icon-bookmark"></use>
-          </svg>
-          <span className="visually-hidden">
-              `${fromOfferPage ? 'In' : 'To'} ${'bookmarks'}
-          </span>
-        </button>
+
+        <OfferCardDetalButton isFavorite={isFavorite} fromOfferPage={fromOfferPage} id={id}/>
+
       </div>
       <div className="place-card__rating rating">
         <div className="place-card__stars rating__stars">
